+++ base/src/main/java/proguard/analysis/cpa/domain/taint/TaintSink.java
@@ -18,10 +18,6 @@
 
 package proguard.analysis.cpa.domain.taint;
 
-import java.util.Objects;
-import java.util.Set;
-import java.util.stream.Collectors;
-
 /**
  * A {@link TaintSink} specifies a sink for the taint analysis. A sink can be sensitive to
  * the instance, the arguments, or the static fields. If a sink S is sensitive to X, then
@@ -29,83 +25,26 @@ import java.util.stream.Collectors;
  *
  * @author Dmitry Ivanov
  */
-public class TaintSink
+public abstract class TaintSink
 {
-    public final String       fqn;
-    public final boolean      takesInstance;
-    public final Set<Integer> takesArgs;
-    public final Set<String>  takesGlobals;
+    public final String fqn;
 
     /**
      * Create a taint sink.
      *
      * @param fqn           the fully qualified name of a sink method
-     * @param takesInstance whether the sink is sensitive to the calling instance
-     * @param takesArgs     a set of sensitive arguments
-     * @param takesGlobals  a set of sensitive global variables
      */
-    public TaintSink(String fqn, boolean takesInstance, Set<Integer> takesArgs, Set<String> takesGlobals)
+    public TaintSink(String fqn)
     {
-        if (!takesInstance && takesArgs.isEmpty() && takesGlobals.isEmpty())
-        {
-            throw new RuntimeException(String.format("Tainted sink for method %s must have taint somewhere!", fqn));
-        }
         this.fqn = fqn;
-        this.takesInstance = takesInstance;
-        this.takesArgs = takesArgs;
-        this.takesGlobals = takesGlobals;
     }
 
     @Override
-    public boolean equals(Object o)
-    {
-        if (this == o)
-        {
-            return true;
-        }
-        if (!(o instanceof TaintSink))
-        {
-            return false;
-        }
-        TaintSink taintSink = (TaintSink) o;
-        return takesInstance == taintSink.takesInstance
-               && Objects.equals(fqn, taintSink.fqn)
-               && Objects.equals(takesArgs, taintSink.takesArgs)
-               && Objects.equals(takesGlobals, taintSink.takesGlobals);
-    }
+    public abstract boolean equals(Object o);
 
     @Override
-    public int hashCode()
-    {
-        return Objects.hash(fqn, takesInstance, takesArgs, takesGlobals);
-    }
+    public abstract int hashCode();
 
     @Override
-    public String toString()
-    {
-        StringBuilder result = new StringBuilder("[TaintSink] ").append(fqn);
-        if (takesInstance)
-        {
-            result.append(", takes instance");
-        }
-        if (!takesArgs.isEmpty())
-        {
-            result.append(", takes args (")
-                  .append(takesArgs.stream()
-                                   .map(Objects::toString)
-                                   .sorted()
-                                   .collect(Collectors.joining(", ")))
-                  .append(")");
-        }
-        if (!takesGlobals.isEmpty())
-        {
-            result.append(", takes globals (")
-                  .append(takesGlobals.stream()
-                                      .sorted()
-                                      .collect(Collectors.joining(", ")))
-                  .append(")");
-        }
-
-        return result.toString();
-    }
+    public abstract String toString();
 }