+++ base/src/test/kotlin/proguard/analysis/cpa/HeapOperatorsTest.kt
@@ -16,7 +16,6 @@ import proguard.analysis.cpa.jvm.domain.reference.CompositeHeapJvmAbstractState
 import proguard.analysis.cpa.jvm.domain.reference.JvmReferenceAbstractState
 import proguard.analysis.cpa.jvm.domain.reference.Reference
 import proguard.analysis.cpa.jvm.domain.taint.JvmTaintMemoryLocationBamCpaRun
-import proguard.analysis.cpa.jvm.domain.taint.JvmTaintSink
 import proguard.analysis.cpa.jvm.state.JvmAbstractState
 import proguard.analysis.cpa.jvm.state.heap.HeapModel
 import proguard.analysis.cpa.jvm.state.heap.tree.HeapNode
@@ -34,6 +33,7 @@ import proguard.testutils.JavaSource
 import java.util.Optional
 import kotlin.reflect.full.memberProperties
 import kotlin.reflect.jvm.isAccessible
+import proguard.analysis.cpa.jvm.domain.taint.JvmInvokeTaintSink
 
 @Ignored
 class HeapOperatorsTest : FreeSpec({
@@ -85,7 +85,7 @@ class HeapOperatorsTest : FreeSpec({
         setOf()
     )
 
-    val taintSinkArgument = JvmTaintSink(
+    val taintSinkArgument = JvmInvokeTaintSink(
         "LA;sink(Ljava/lang/String;)V",
         false,
         setOf(1),