+++ base/src/test/kotlin/proguard/analysis/cpa/TraceExtractorTest.kt
@@ -23,7 +23,6 @@ import io.kotest.matchers.shouldBe
 import proguard.analysis.cpa.domain.taint.TaintAbstractState
 import proguard.analysis.cpa.domain.taint.TaintSource
 import proguard.analysis.cpa.jvm.domain.taint.JvmTaintMemoryLocationBamCpaRun
-import proguard.analysis.cpa.jvm.domain.taint.JvmTaintSink
 import proguard.analysis.cpa.jvm.util.CfaUtil
 import proguard.analysis.cpa.state.DifferentialMapAbstractStateFactory
 import proguard.analysis.cpa.state.HashMapAbstractStateFactory
@@ -31,6 +30,8 @@ import proguard.analysis.cpa.state.LimitedHashMapAbstractStateFactory
 import proguard.testutils.ClassPoolBuilder
 import proguard.testutils.JavaSource
 import java.util.Optional
+import proguard.analysis.cpa.jvm.domain.taint.JvmInvokeTaintSink
+import proguard.analysis.cpa.jvm.domain.taint.JvmReturnTaintSink
 
 class TraceExtractorTest : StringSpec({
 
@@ -56,30 +57,31 @@ class TraceExtractorTest : StringSpec({
         setOf("A.s")
     )
 
-    val taintSinkArgument = JvmTaintSink(
+    val taintSinkArgument = JvmInvokeTaintSink(
         "LA;sink(Ljava/lang/String;)V",
         false,
         setOf(1),
         setOf()
     )
-    val taintSinkArgumentLong = JvmTaintSink(
+    val taintSinkArgumentLong = JvmInvokeTaintSink(
         "LA;sink(J)V",
         false,
         setOf(1),
         setOf()
     )
-    val taintSinkArgumentMultiple = JvmTaintSink(
+    val taintSinkArgumentMultiple = JvmInvokeTaintSink(
         "LA;sink(Ljava/lang/String;Ljava/lang/String;)V",
         false,
         setOf(2),
         setOf()
     )
-    val taintSinkStatic = JvmTaintSink(
+    val taintSinkStatic = JvmInvokeTaintSink(
         "LA;sink()V",
         false,
         setOf(),
         setOf("A.s")
     )
+    val taintSinkReturn = JvmReturnTaintSink("LA;sink(Ljava/lang/String;)Ljava/lang/String;")
 
     val jvmTaintMemoryLocationBamCpaRunBuilder = JvmTaintMemoryLocationBamCpaRun.Builder().setReduceHeap(false)
 
@@ -1129,6 +1131,73 @@ class TraceExtractorTest : StringSpec({
                     "JvmStaticFieldLocation(A.s)@LA;main()V:3",
                     "JvmStaticFieldLocation(A.s)@LA;callee()V:6",
                     "JvmStackLocation(0)@LA;callee()V:3"
+                )
+            )
+        }
+
+        "Simple interprocedural flows with return sink are reconstructed$testNameSuffix" {
+            val interproceduralCfa = CfaUtil.createInterproceduralCfaFromClassPool(
+                ClassPoolBuilder.fromSource(
+                    JavaSource(
+                        "A.java",
+                        """
+                    class A {
+
+                        public void main() {
+                            sink(callee());
+                        }
+                    
+                        public static String callee()
+                        {
+                            return source1();
+                        }
+                    
+                        public static String sink(String s)
+                        {
+                            return s;
+                        }
+                    
+                        public static String source1()
+                        {
+                            return null;
+                        }
+                    }
+                        """.trimIndent()
+                    ),
+                    javacArguments = listOf("-source", "1.8", "-target", "1.8")
+                ).programClassPool
+            )
+            /*
+            Bytecode of main:
+                [0] invokestatic #2 = Methodref(A.callee()Ljava/lang/String;)
+                [3] invokestatic #3 = Methodref(A.sink(Ljava/lang/String;)Ljava/lang/String;)
+                [6] pop
+                [7] return
+
+            Bytecode of callee:
+                [0] invokestatic #4 = Methodref(A.source1()Ljava/lang/String;)
+                [3] areturn
+
+            Bytecode of sink:
+                [0] aload_0 v0
+                [1] areturn
+             */
+            val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
+            val taintMemoryLocationCpaRun = jvmTaintMemoryLocationBamCpaRunBuilder
+                .setCfa(interproceduralCfa)
+                .setMainSignature(mainSignature)
+                .setTaintSources(setOf(taintSourceReturn1))
+                .setTaintSinks(setOf(taintSinkReturn))
+                .build()
+            val traces = taintMemoryLocationCpaRun.extractLinearTraces()
+            interproceduralCfa.clear()
+
+            traces.map { trace -> trace.map { it.toString() } }.toSet() shouldBe setOf(
+                listOf(
+                    "JvmStackLocation(0)@LA;sink(Ljava/lang/String;)Ljava/lang/String;:1",
+                    "JvmLocalVariableLocation(0)@LA;sink(Ljava/lang/String;)Ljava/lang/String;:0",
+                    "JvmStackLocation(0)@LA;main()V:3",
+                    "JvmStackLocation(0)@LA;callee()Ljava/lang/String;:3"
                 )
             )
         }